src/data_indexer/requirements.txt
chromadb==1.0.13
cppyy==3.5.0
graphviz==0.21
langchain==0.3.26
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.0
psutil==7.0.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
ROOT==0.0.1
scipy==1.16.0
seaborn==0.13.2
sentence_transformers==4.1.0
torch==2.7.0
